travel_pipeline/pyproject.toml
[project]
name = "travel-pipeline"
version = "0.1.0"
description = "Distributed BTS flight pipeline on MongoDB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"pandas>=2.2",
	"pymongo>=4.7",
	"pydantic>=2.6",
	"python-dotenv>=1.0",
	"loguru>=0.7",
	"streamlit>=1.39",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.3",
	"pytest-cov>=4.1",
	"mypy>=1.11",
]

[project.scripts]
travel-pipeline-ingest = "travel_pipeline.cli:run_ingest"
travel-pipeline-clean = "travel_pipeline.cli:run_clean"
travel-pipeline-aggregate = "travel_pipeline.cli:run_aggregate"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
